+++ r10k_modules/windows_firewall/lib/puppet/provider/windows_firewall_group/ruby.rb
@@ -1,13 +1,12 @@
 require 'puppet_x'
-#require 'puppet_x/windows_firewall'
 require_relative '../../../puppet_x/windows_firewall'
 
 Puppet::Type.type(:windows_firewall_group).provide(:windows_firewall_group, :parent => Puppet::Provider) do
   confine :osfamily => :windows
   mk_resource_methods
-  desc "Windows Firewall group"
+  desc 'Windows Firewall group'
 
-  commands :cmd => "netsh"
+  commands :cmd => 'netsh'
 
   def self.prefetch(resources)
     instances.each do |prov|
@@ -19,18 +18,14 @@ Puppet::Type.type(:windows_firewall_group).provide(:windows_firewall_group, :par
 
   # firewall groups always exist we can only enable/disable them
   def exists?
-    #@property_hash[:ensure] == :present
     true
   end
 
   # all work done in `flush()` method
-  def create()
-  end
+  def create; end
 
   # all work done in `flush()` method
-  def destroy()
-  end
-
+  def destroy; end
 
   def self.instances
     PuppetX::WindowsFirewall.groups.collect { |hash| new(hash) }
@@ -42,7 +37,7 @@ Puppet::Type.type(:windows_firewall_group).provide(:windows_firewall_group, :par
     # to inspect @resource instead
 
     # careful its a label not a boolean...
-    netsh_enabled = (@resource[:enabled] == :true)? "yes" : "no"
+    netsh_enabled = (@resource[:enabled] == :true)? 'yes': 'no'
 
     Puppet.notice("(windows_firewall) group '#{@resource[:name]}' enabled: #{@resource[:enabled]}")
     cmd = "#{command(:cmd)} advfirewall firewall set rule group=\"#{@resource[:name]}\" new enable=\"#{netsh_enabled}\""