+++ r10k_modules/windows_firewall/lib/puppet/provider/windows_firewall_rule/ruby.rb
@@ -4,8 +4,7 @@ require_relative '../../../puppet_x/windows_firewall'
 Puppet::Type.type(:windows_firewall_rule).provide(:windows_firewall_rule, :parent => Puppet::Provider) do
   confine :osfamily => :windows
   mk_resource_methods
-  desc "Windows Firewall"
-
+  desc 'Windows Firewall'
 
   def self.prefetch(resources)
     instances.each do |prov|
@@ -14,18 +13,17 @@ Puppet::Type.type(:windows_firewall_rule).provide(:windows_firewall_rule, :paren
       end
     end
   end
-  
+
   def exists?
     @property_hash[:ensure] == :present
   end
 
-  def create()
+  def create
     PuppetX::WindowsFirewall.create_rule @resource
   end
 
-  def destroy()
-    # Delete only require firewall rule name to process
-    PuppetX::WindowsFirewall.delete_rule @resource[:name]
+  def destroy
+    PuppetX::WindowsFirewall.delete_rule @property_hash
   end
 
   def self.instances