+++ r10k_modules/windows_firewall/lib/puppet_x/windows_firewall_ipsec.rb
@@ -3,10 +3,9 @@ require 'pp'
 module PuppetX
   module WindowsFirewallIPSec
 
-    MOD_DIR = "windows_firewall/lib"
-    SCRIPT_FILE = "ps-bridge-ipsec.ps1"
-    SCRIPT_PATH = File.join("ps/windows_firewall", SCRIPT_FILE)
-
+    MOD_DIR = 'windows_firewall/lib'
+    SCRIPT_FILE = 'ps-bridge-ipsec.ps1'
+    SCRIPT_PATH = File.join('ps/windows_firewall', SCRIPT_FILE)
 
     # We need to be able to invoke the PS bridge script in both agent and apply
     # mode. In agent mode, the file will be found in LIBDIR, in apply mode it will
@@ -32,7 +31,7 @@ module PuppetX
         raise("windows_firewall unable to find #{SCRIPT_FILE} in expected location")
       end
 
-      cmd = ["powershell.exe", "-ExecutionPolicy", "Bypass", "-File", script]
+      cmd = ['powershell.exe', '-ExecutionPolicy', 'Bypass', '-File', script]
       cmd
     end
 
@@ -41,7 +40,7 @@ module PuppetX
       check_for_script = File.join(
           Puppet.settings[:environmentpath],
           Puppet.settings[:environment],
-          "modules",
+          'modules',
           MOD_DIR,
           SCRIPT_PATH,
           )
@@ -76,30 +75,31 @@ module PuppetX
 
     def self.to_ps(key)
       {
-        :enabled               => lambda { |x| camel_case(x)},
-        :action                => lambda { |x| camel_case(x)},
-        :interface_type        => lambda { |x| x.map {|e| camel_case(e)}.join(",")},
-        :profile               => lambda { |x| x.map {|e| camel_case(e)}.join(",")},
-        :protocol              => lambda { |x| x.to_s.upcase.sub("V","v")},
-        :local_port            => lambda { |x| "\"#{camel_case(x)}\""},
-        :remote_port           => lambda { |x| "\"#{camel_case(x)}\""},
-        :local_address         => lambda { |x| "\"#{camel_case(x)}\""},
-        :remote_address        => lambda { |x| "\"#{camel_case(x)}\""},
-        :mode                  => lambda { |x| camel_case(x)},
-        :inbound_security      => lambda { |x| camel_case(x)},
-        :outbound_security     => lambda { |x| camel_case(x)},
-        :phase1auth_set        => lambda { |x| camel_case(x)},
-        :phase2auth_set        => lambda { |x| camel_case(x)},
+        :enabled               => lambda { |x| camel_case(x) },
+        :action                => lambda { |x| camel_case(x) },
+        :description           => lambda { |x| x.empty? == true ? "\"#{x}\"" : x },
+        :interface_type        => lambda { |x| x.map { |e| camel_case(e)}.join(',') },
+        :profile               => lambda { |x| x.map { |e| camel_case(e)}.join(',') },
+        :protocol              => lambda { |x| x.to_s.upcase.sub('V', 'v') },
+        :local_port            => lambda { |x| "\"#{camel_case(x)}\"" },
+        :remote_port           => lambda { |x| "\"#{camel_case(x)}\"" },
+        :local_address         => lambda { |x| "\"#{camel_case(x)}\"" },
+        :remote_address        => lambda { |x| "\"#{camel_case(x)}\"" },
+        :mode                  => lambda { |x| camel_case(x) },
+        :inbound_security      => lambda { |x| camel_case(x) },
+        :outbound_security     => lambda { |x| camel_case(x) },
+        :phase1auth_set        => lambda { |x| camel_case(x) },
+        :phase2auth_set        => lambda { |x| camel_case(x) },
       }.fetch(key, lambda { |x| x })
     end
 
     def self.to_ruby(key)
       {
-        :enabled                => lambda { |x| snake_case_sym(x)},
-        :action                 => lambda { |x| snake_case_sym(x)},
-        :interface_type         => lambda { |x| x.split(",").map{ |e| snake_case_sym(e.strip)}},
-        :profile                => lambda { |x| x.split(",").map{ |e| snake_case_sym(e.strip)}},
-        :protocol               => lambda { |x| snake_case_sym(x)},
+        :enabled                => lambda { |x| snake_case_sym(x) },
+        :action                 => lambda { |x| snake_case_sym(x) },
+        :interface_type         => lambda { |x| x.split(',').map{ |e| snake_case_sym(e.strip) } },
+        :profile                => lambda { |x| x.split(',').map{ |e| snake_case_sym(e.strip) } },
+        :protocol               => lambda { |x| snake_case_sym(x) },
         :remote_port            => lambda { |x| x.downcase },
         :local_port             => lambda { |x| x.downcase },
         :remote_address         => lambda { |x| x.downcase },
@@ -117,7 +117,7 @@ module PuppetX
     # 2. converting spaces to underscores
     # 3. convert to symbol
     def self.key_name(input)
-      input.downcase.gsub(/\s/, "_").to_sym
+      input.downcase.gsub(/\s/, '_').to_sym
     end
 
     # Convert input CamelCase to snake_case symbols
@@ -131,21 +131,20 @@ module PuppetX
       input.to_s.split('_').collect(&:capitalize).join
     end
 
-    def self.delete_rule(name)
-      Puppet.notice("(windows_firewall) deleting rule '#{name}'")
-      out = Puppet::Util::Execution.execute(resolve_ps_bridge + ["delete", name]).to_s
+    def self.delete_rule(resource)
+      Puppet.notice("(windows_firewall) deleting ipsec rule '#{resource[:display_name]}'")
+      out = Puppet::Util::Execution.execute(resolve_ps_bridge + ['delete', resource[:name]]).to_s
       Puppet.debug out
     end
 
     def self.update_rule(resource)
-      Puppet.notice("(windows_firewall) updating rule '#{resource[:name]}'")
+      Puppet.notice("(windows_firewall) updating ipsec rule '#{resource[:display_name]}'")
 
       # `Name` is mandatory and also a `parameter` not a `property`
-      args = [ "-Name", resource[:name] ]
-      
+      args = [ '-Name', resource[:name] ]
+
       resource.properties.reject { |property|
-        [:ensure, :protocol_type, :protocol_code].include?(property.name) ||
-            property.value == :none
+        [:ensure, :protocol_type, :protocol_code].include?(property.name)
       }.each { |property|
         # All properties start `-`
         property_name = "-#{camel_case(property.name)}"
@@ -155,23 +154,22 @@ module PuppetX
         args << property_name
         args << property_value
       }
-      Puppet.debug "Updating firewall rule with args: #{args}"
+      Puppet.debug "Updating firewall ipsec rule with args: #{args}"
 
-      out = Puppet::Util::Execution.execute(resolve_ps_bridge + ["update"] + args)
+      out = Puppet::Util::Execution.execute(resolve_ps_bridge + ['update'] + args)
       Puppet.debug out
     end
 
     # Create a new firewall rule using powershell
     # @see https://docs.microsoft.com/en-us/powershell/module/netsecurity/new-netfirewallrule?view=win10-ps
     def self.create_rule(resource)
-      Puppet.notice("(windows_firewall) adding rule '#{resource[:name]}'")
+      Puppet.notice("(windows_firewall) adding ipsec rule '#{resource[:display_name]}'")
 
       # `Name` is mandatory and also a `parameter` not a `property`
-      args = [ "-Name", resource[:name] ]
-      
+      args = [ '-Name', resource[:name] ]
+
       resource.properties.reject { |property|
-        [:ensure, :protocol_type, :protocol_code].include?(property.name) ||
-            property.value == :none
+        [:ensure, :protocol_type, :protocol_code].include?(property.name)
       }.each { |property|
         # All properties start `-`
         property_name = "-#{camel_case(property.name)}"
@@ -181,25 +179,25 @@ module PuppetX
         args << property_name
         args << property_value
       }
-      Puppet.debug "Creating firewall rule with args: #{args}"
+      Puppet.debug "Creating firewall ipsec rule with args: #{args}"
 
-      out = Puppet::Util::Execution.execute(resolve_ps_bridge + ["create"] + args)
+      out = Puppet::Util::Execution.execute(resolve_ps_bridge + ['create'] + args)
       Puppet.debug out
     end
 
     def self.rules
-      Puppet.debug("query all rules")
-      rules = JSON.parse Puppet::Util::Execution.execute(resolve_ps_bridge + ["show"]).to_s
+      Puppet.debug('query all ipsec rules')
+      rules = JSON.parse Puppet::Util::Execution.execute(resolve_ps_bridge + ['show']).to_s
 
       # Rules is an array of hash as-parsed and hash keys need converted to
       # lowercase ruby labels
       puppet_rules = rules.map { |e|
-        Hash[e.map { |k,v|
+        Hash[e.map { |k, v|
           key = snake_case_sym(k)
           [key, to_ruby(key).call(v)]
         }].merge({ensure: :present})
       }
-      Puppet.debug("Parsed rules: #{puppet_rules.size}")
+      Puppet.debug("Parsed ipsec rules: #{puppet_rules.size}")
       puppet_rules
     end
 
@@ -207,7 +205,7 @@ module PuppetX
     def self.parse_profile(input)
       profile = {}
       first_line = true
-      profile_name = "__error__"
+      profile_name = '__error__'
       input.split("\n").reject { |line|
         line =~ /---/ || line =~ /^\s*$/
       }.each { |line|
@@ -218,7 +216,7 @@ module PuppetX
         else
           # nasty hack - "firewall policy" setting contains space and will break our
           # logic below. Also the setter in `netsh` to use is `firewallpolicy`. Just fix it...
-          line = line.sub("Firewall Policy", "firewallpolicy")
+          line = line.sub('Firewall Policy', 'firewallpolicy')
 
           # split each line at most twice by first glob of whitespace
           line_split = line.split(/\s+/, 2)
@@ -263,17 +261,17 @@ module PuppetX
             #   DHGroup2-AES128-SHA1,DHGroup2-3DES-SHA1
             # but must be input with a colon like this:
             #   DHGroup2:AES128-SHA1,DHGroup2:3DES-SHA1
-            safe_value = value.split(",").map { |e|
-              e.sub("-", ":")
-            }.join(",")
+            safe_value = value.split(',').map { |e|
+              e.sub('-', ':')
+            }.join(',')
           when :strongcrlcheck
-            safe_value = value.split(":")[0]
+            safe_value = value.split(':')[0]
           when :defaultexemptions
-            safe_value = value.split(",").sort
+            safe_value = value.split(',').sort
           when :saidletimemin
-            safe_value = value.sub("min","")
+            safe_value = value.sub('min', '')
           when :ipsecthroughnat
-            safe_value = value.gsub(" ","")
+            safe_value = value.gsub(' ', '')
           else
             safe_value = value
           end
@@ -282,7 +280,7 @@ module PuppetX
         end
       }
 
-      globals[:name] = "global"
+      globals[:name] = 'global'
 
       Puppet.debug "Parsed windows firewall globals: #{globals}"
       globals