+++ r10k_modules/windows_firewall/CHANGELOG.md
@@ -2,6 +2,74 @@
 
 All notable changes to this project will be documented in this file.
 
+## Release 1.3.4 (2022-08-23)
+
+[Full Changelog](https://github.com/webalexeu/puppet-windows_firewall/compare/v1.3.3...v1.3.4)
+
+**Features**
+
+ - Add support for Windows Server 2022
+ - Add support for Puppet 7
+ - Code cleaning
+
+**Bugfixes**
+
+**Known Issues**
+
+## Release 1.3.3 (2022-08-20)
+
+[Full Changelog](https://github.com/webalexeu/puppet-windows_firewall/compare/v1.3.2...v1.3.3)
+
+**Features**
+
+**Bugfixes**
+
+- [Errors with windows_firewall_group resource](https://github.com/webalexeu/puppet-windows_firewall/issues/21)
+
+**Known Issues**
+
+## Release 1.3.2 (2022-08-16)
+
+[Full Changelog](https://github.com/webalexeu/puppet-windows_firewall/compare/v1.3.1...v1.3.2)
+
+**Features**
+
+- Change execution output from name to display_name (user-facing name). Default rules name are a randomly assigned value by default (Ex: {F207584F-6202-41D0-B097-6C232F8B64CD}). When you are using purge mechanism, deletion of default rules output will be more user-friendly
+
+**Bugfixes**
+
+**Known Issues**
+
+- Errors with windows_firewall_group resource
+
+## Release 1.3.1 (2022-08-02)
+
+[Full Changelog](https://github.com/webalexeu/puppet-windows_firewall/compare/v1.3.0...v1.3.1)
+
+**Features**
+
+**Bugfixes**
+
+- [Corrective action on local_port when protocol is icmpv4 and icmp_type is not any](https://github.com/webalexeu/puppet-windows_firewall/issues/17)
+- [Corrective action for some description attribute](https://github.com/webalexeu/puppet-windows_firewall/issues/18)
+
+**Known Issues**
+
+## Release 1.3.0 (2022-07-29)
+
+[Full Changelog](https://github.com/webalexeu/puppet-windows_firewall/compare/v1.2.3...v1.3.0)
+
+**Features**
+
+- Manage default value. If no value is specified (for optional settings), the default will be set. This will also ensure that all settings are maintained by the module
+
+**Bugfixes**
+
+**Known Issues**
+
+- Corrective action on local_port when protocol is icmpv4 and icmp_type is not any
+- Corrective action for some description attribute
+
 ## Release 1.2.3 (2021-06-09)
 
 [Full Changelog](https://github.com/webalexeu/puppet-windows_firewall/compare/v1.2.2...v1.2.3)